split bump collider into separate component
