+++ components/movement/circle_mover.tscn
@@ -1,0 +1,6 @@
+[gd_scene load_steps=2 format=3 uid="uid://dncccyyoapb1v"]
+
+[ext_resource type="Script" path="res://components/movement/CircleMover.cs" id="1_3wxc4"]
+
+[node name="CircleMover" type="Node"]
+script = ExtResource("1_3wxc4")
