--- components/movement/path_stepper.tscn
@@ -1,6 +1,0 @@
-[gd_scene load_steps=2 format=3 uid="uid://wa4p4jkrefpa"]
-
-[sub_resource type="Curve2D" id="Curve2D_ydsqy"]
-
-[node name="PathStepper" type="Path2D"]
-curve = SubResource("Curve2D_ydsqy")
