+++ components/movement/velocity_mover.tscn
@@ -1,16 +1,6 @@
-[gd_scene load_steps=3 format=3 uid="uid://d1qe8j2vcl8xn"]
+[gd_scene load_steps=2 format=3 uid="uid://d1qe8j2vcl8xn"]
 
 [ext_resource type="Script" path="res://components/movement/VelocityMover.cs" id="1_3c6ne"]
 
-[sub_resource type="CircleShape2D" id="CircleShape2D_xq56w"]
-radius = 1.0
-
-[node name="VelocityMover" type="Area2D"]
-collision_layer = 4
-collision_mask = 4
+[node name="VelocityMover" type="Node"]
 script = ExtResource("1_3c6ne")
-
-[node name="CollisionShape2D" type="CollisionShape2D" parent="."]
-shape = SubResource("CircleShape2D_xq56w")
-
-[connection signal="area_entered" from="." to="." method="OnAreaEntered"]
