+++ scenes/test.tscn
@@ -1,16 +1,16 @@
 [gd_scene load_steps=4 format=3 uid="uid://c0ybtmwx2a0h5"]
 
 [ext_resource type="PackedScene" uid="uid://bbwuhq5a4pofj" path="res://scenes/background.tscn" id="1_5vlkb"]
-[ext_resource type="PackedScene" uid="uid://dpswa70jplyuu" path="res://entities/actors/player/player.tscn" id="2_jljom"]
-[ext_resource type="PackedScene" uid="uid://4de4nosgm5gn" path="res://entities/actors/enemy/straight_line_enemy.tscn" id="3_1io6m"]
+[ext_resource type="PackedScene" uid="uid://b6vf7xrp51cib" path="res://entities/actors/enemy/base_enemy.tscn" id="2_355aj"]
+[ext_resource type="PackedScene" uid="uid://dncccyyoapb1v" path="res://components/movement/circle_mover.tscn" id="3_5sqqy"]
 
 [node name="Test" type="Node2D"]
 
 [node name="Background" parent="." instance=ExtResource("1_5vlkb")]
 
-[node name="Player" parent="." instance=ExtResource("2_jljom")]
-position = Vector2(192, 192)
+[node name="BaseEnemy" parent="." instance=ExtResource("2_355aj")]
+position = Vector2(448, 320)
 
-[node name="StraightLineEnemy2" parent="." instance=ExtResource("3_1io6m")]
-position = Vector2(320, 320)
-rotation = 1.5708
+[node name="CircleMover" parent="BaseEnemy" node_paths=PackedStringArray("Body") instance=ExtResource("3_5sqqy")]
+RotationsPerBeat = 0.25
+Body = NodePath("..")
